@@ -1,19 +1,18 @@
 name: myanmar_calendar_dart
 description: A comprehensive dart package for Myanmar calendar with date conversions, astrological calculations, and multi-language support
-version: 1.1.5
+version: 2.0.0
 homepage: https://github.com/mixin27/myanmar_calendar_dart
 repository: https://github.com/mixin27/myanmar_calendar_dart
 issue_tracker: https://github.com/mixin27/myanmar_calendar_dart/issues
 
 environment:
-  sdk: ^3.9.0
+  sdk: ^3.11.0
 
 dependencies:
   collection: ^1.19.1
   meta: ^1.17.0
 
 dev_dependencies:
-  mocktail: ^1.0.4
   test: ^1.26.3
   very_good_analysis: ^10.0.0
 
