@@ -1,6 +1,7 @@
 import 'package:collection/collection.dart';
 import 'package:myanmar_calendar_dart/src/models/custom_holiday.dart';
 import 'package:myanmar_calendar_dart/src/models/holiday_id.dart';
+import 'package:myanmar_calendar_dart/src/models/western_holiday_provider.dart';
 
 /// Configuration class for the Myanmar Calendar
 class CalendarConfig {
@@ -15,7 +16,21 @@ class CalendarConfig {
     this.disabledHolidays = const [],
     this.disabledHolidaysByYear = const {},
     this.disabledHolidaysByDate = const {},
-  });
+    this.westernHolidayProvider = const DefaultWesternHolidayProvider(),
+  }) : assert(
+         sasanaYearType >= 0 && sasanaYearType <= 2,
+         'sasanaYearType must be between 0 and 2',
+       ),
+       assert(
+         calendarType >= 0 && calendarType <= 2,
+         'calendarType must be between 0 and 2',
+       ),
+       assert(
+         timezoneOffset >= -12 && timezoneOffset <= 14,
+         'timezoneOffset must be between -12 and 14',
+       ),
+       assert(gregorianStart > 0, 'gregorianStart must be greater than 0'),
+       assert(defaultLanguage != '', 'defaultLanguage must not be empty');
 
   /// Create config with Myanmar Time (UTC+6:30)
   factory CalendarConfig.myanmarTime() {
@@ -69,6 +84,11 @@ class CalendarConfig {
   /// Custom holidays defined by the consumer
   final List<CustomHoliday> customHolidays;
 
+  /// Custom holiday rules defined by the consumer.
+  ///
+  /// This is the preferred naming over [customHolidays].
+  List<CustomHoliday> get customHolidayRules => customHolidays;
+
   /// List of built-in holidays to disable globally
   final List<HolidayId> disabledHolidays;
 
@@ -77,6 +97,9 @@ class CalendarConfig {
 
   /// Map of Western date (YYYY-MM-DD) to list of built-in holidays to disable for that specific date
   final Map<String, List<HolidayId>> disabledHolidaysByDate;
+
+  /// Provider for western-calendar holiday rules (e.g., Eid/Diwali/CNY).
+  final WesternHolidayProvider westernHolidayProvider;
 
   /// Get the current timezone offset in days
   double get timezoneOffsetInDays => timezoneOffset / 24.0;
@@ -87,6 +110,48 @@ class CalendarConfig {
   /// Convert UTC Julian Day Number to local time
   double utcToLocal(double utcJdn) => utcJdn + timezoneOffsetInDays;
 
+  /// Cache namespace fingerprint for this configuration.
+  ///
+  /// This is used to isolate cache entries across different configurations
+  /// when sharing a global cache instance.
+  String get cacheNamespace {
+    final globalDisabled = disabledHolidays.map((e) => e.name).toList()..sort();
+
+    final yearSpecific = disabledHolidaysByYear.entries.toList()
+      ..sort((a, b) => a.key.compareTo(b.key));
+    final yearSpecificKey = yearSpecific
+        .map((entry) {
+          final ids = entry.value.map((e) => e.name).toList()..sort();
+          return '${entry.key}:${ids.join('.')}';
+        })
+        .join('|');
+
+    final dateSpecific = disabledHolidaysByDate.entries.toList()
+      ..sort((a, b) => a.key.compareTo(b.key));
+    final dateSpecificKey = dateSpecific
+        .map((entry) {
+          final ids = entry.value.map((e) => e.name).toList()..sort();
+          return '${entry.key}:${ids.join('.')}';
+        })
+        .join('|');
+
+    final customHolidayKey = customHolidays.map((holiday) {
+      return holiday.cacheDescriptor;
+    }).toList()..sort();
+    final customHolidayFingerprint = customHolidayKey.join('|');
+
+    return 'sy:$sasanaYearType'
+        '|ct:$calendarType'
+        '|gs:$gregorianStart'
+        '|tz:${timezoneOffset.toStringAsFixed(6)}'
+        '|lang:$defaultLanguage'
+        '|dg:${globalDisabled.join(',')}'
+        '|dy:$yearSpecificKey'
+        '|dd:$dateSpecificKey'
+        '|ch:$customHolidayFingerprint'
+        '|whp:${westernHolidayProvider.cacheKey}';
+  }
+
   /// Copy with new values
   CalendarConfig copyWith({
     int? sasanaYearType,
@@ -94,10 +159,11 @@ class CalendarConfig {
     int? gregorianStart,
     double? timezoneOffset,
     String? defaultLanguage,
-    List<CustomHoliday>? customHolidays,
+    List<CustomHoliday>? customHolidayRules,
     List<HolidayId>? disabledHolidays,
     Map<int, List<HolidayId>>? disabledHolidaysByYear,
     Map<String, List<HolidayId>>? disabledHolidaysByDate,
+    WesternHolidayProvider? westernHolidayProvider,
   }) {
     return CalendarConfig(
       sasanaYearType: sasanaYearType ?? this.sasanaYearType,
@@ -105,12 +171,14 @@ class CalendarConfig {
       gregorianStart: gregorianStart ?? this.gregorianStart,
       timezoneOffset: timezoneOffset ?? this.timezoneOffset,
       defaultLanguage: defaultLanguage ?? this.defaultLanguage,
-      customHolidays: customHolidays ?? this.customHolidays,
+      customHolidays: customHolidayRules ?? customHolidays,
       disabledHolidays: disabledHolidays ?? this.disabledHolidays,
       disabledHolidaysByYear:
           disabledHolidaysByYear ?? this.disabledHolidaysByYear,
       disabledHolidaysByDate:
           disabledHolidaysByDate ?? this.disabledHolidaysByDate,
+      westernHolidayProvider:
+          westernHolidayProvider ?? this.westernHolidayProvider,
     );
   }
 
@@ -139,7 +207,9 @@ class CalendarConfig {
         const MapEquality<String, List<HolidayId>>().equals(
           other.disabledHolidaysByDate,
           disabledHolidaysByDate,
-        );
+        ) &&
+        other.westernHolidayProvider.cacheKey ==
+            westernHolidayProvider.cacheKey;
   }
 
   @override
@@ -154,6 +224,7 @@ class CalendarConfig {
         const MapEquality<int, List<HolidayId>>().hash(disabledHolidaysByYear) ^
         const MapEquality<String, List<HolidayId>>().hash(
           disabledHolidaysByDate,
-        );
+        ) ^
+        westernHolidayProvider.cacheKey.hashCode;
   }
 }
