@@ -16,7 +16,8 @@ enum Language {
   shan('shan'),
 
   /// Karen
-  karen('karen');
+  karen('karen')
+  ;
 
   const Language(this.code);
 
