@@ -1,8 +1,8 @@
 /// Contains astronomical and astrological information for a Myanmar date
 class AstroInfo {
   /// Create a new astro info
-  const AstroInfo({
-    required this.astrologicalDays,
+  AstroInfo({
+    required List<String> astrologicalDays,
     required this.sabbath,
     required this.yatyaza,
     required this.pyathada,
@@ -10,7 +10,7 @@ class AstroInfo {
     required this.mahabote,
     required this.nakhat,
     required this.yearName,
-  });
+  }) : astrologicalDays = List.unmodifiable(astrologicalDays);
 
   /// Astrological days
   final List<String> astrologicalDays;
@@ -60,5 +60,39 @@ class AstroInfo {
         'sabbathInfo: $sabbath, yatyaza: $yatyaza, pyathada: $pyathada, '
         'nagahle: $nagahle, mahabote: $mahabote, nakhat: $nakhat, '
         'yearName: $yearName)';
+  }
+
+  @override
+  bool operator ==(Object other) {
+    if (identical(this, other)) return true;
+    return other is AstroInfo &&
+        other.sabbath == sabbath &&
+        other.yatyaza == yatyaza &&
+        other.pyathada == pyathada &&
+        other.nagahle == nagahle &&
+        other.mahabote == mahabote &&
+        other.nakhat == nakhat &&
+        other.yearName == yearName &&
+        _listEquals(other.astrologicalDays, astrologicalDays);
+  }
+
+  @override
+  int get hashCode => Object.hash(
+    Object.hashAll(astrologicalDays),
+    sabbath,
+    yatyaza,
+    pyathada,
+    nagahle,
+    mahabote,
+    nakhat,
+    yearName,
+  );
+
+  static bool _listEquals(List<String> a, List<String> b) {
+    if (a.length != b.length) return false;
+    for (var i = 0; i < a.length; i++) {
+      if (a[i] != b[i]) return false;
+    }
+    return true;
   }
 }
