@@ -3,17 +3,19 @@ import 'package:myanmar_calendar_dart/src/localization/translation_service.dart'
 
 /// Descriptive information for Myanmar astrological elements
 class AstroDetails {
+  static const Language _fallbackLanguage = Language.english;
+
   /// Get description for Nakhat type
   static String getNakhatDescription(String nakhat, {Language? language}) {
+    final currentLanguage = language ?? _fallbackLanguage;
     final key = 'desc_${nakhat.toLowerCase()}';
     if (TranslationService.hasTranslation(key)) {
-      return language != null
-          ? TranslationService.translateTo(key, language)
-          : TranslationService.translate(key);
+      return TranslationService.translateTo(key, currentLanguage);
     }
-    return language != null
-        ? TranslationService.translateTo('desc_not_available', language)
-        : TranslationService.translate('desc_not_available');
+    return TranslationService.translateTo(
+      'desc_not_available',
+      currentLanguage,
+    );
   }
 
   /// Get characteristics for Mahabote type
@@ -21,15 +23,15 @@ class AstroDetails {
     String mahabote, {
     Language? language,
   }) {
+    final currentLanguage = language ?? _fallbackLanguage;
     final key = 'desc_${mahabote.toLowerCase()}';
     if (TranslationService.hasTranslation(key)) {
-      return language != null
-          ? TranslationService.translateTo(key, language)
-          : TranslationService.translate(key);
+      return TranslationService.translateTo(key, currentLanguage);
     }
-    return language != null
-        ? TranslationService.translateTo('desc_not_available', language)
-        : TranslationService.translate('desc_not_available');
+    return TranslationService.translateTo(
+      'desc_not_available',
+      currentLanguage,
+    );
   }
 
   /// Get description for astrological days
@@ -37,14 +39,14 @@ class AstroDetails {
     String dayName, {
     Language? language,
   }) {
+    final currentLanguage = language ?? _fallbackLanguage;
     final key = 'desc_${dayName.toLowerCase()}';
     if (TranslationService.hasTranslation(key)) {
-      return language != null
-          ? TranslationService.translateTo(key, language)
-          : TranslationService.translate(key);
+      return TranslationService.translateTo(key, currentLanguage);
     }
-    return language != null
-        ? TranslationService.translateTo('desc_not_available', language)
-        : TranslationService.translate('desc_not_available');
+    return TranslationService.translateTo(
+      'desc_not_available',
+      currentLanguage,
+    );
   }
 }
